requirements.txt
slack-machine==0.18
